requirements.txt
python-telegram-bot==13.7
pandas==1.3.0
